Test categoryMixin and rename other mixins for easier identification
